[MIG] sale_commission_partial_settlement: Migration to 16.0
(cherry picked from commit 3eca214a62461f056e3284ecb296b7edd1ef255e)
