+++ sale_commission_partial_settlement/__manifest__.py
@@ -1,7 +1,7 @@
 # Copyright 2023 Nextev
 {
     "name": "Sales commissions based on paid amount",
-    "version": "14.0.1.0.1",
+    "version": "16.0.1.0.0",
     "author": "Nextev Srl," "Ooops," "Odoo Community Association (OCA)",
     "maintainers": ["aleuffre", "renda-dev", "PicchiSeba"],
     "category": "Sales Management",
@@ -10,7 +10,7 @@
     "website": "https://github.com/OCA/commission",
     "data": [
         "security/ir.model.access.csv",
-        "views/sale_commission_view.xml",
+        "views/commission_view.xml",
     ],
     "installable": True,
 }
