+++ sale_commission_partial_settlement/models/__init__.py
@@ -1,4 +1,6 @@
-from . import sale_commission
-from . import account_move
+from . import commission
+from . import account_invoice_line_agent
+from . import account_invoice_line_agent_partial
 from . import account_partial_reconcile
-from . import settlement
+from . import commission_settlement
+from . import commission_settlement_line
