+++ sale_commission_partial_settlement/models/commission_settlement.py
@@ -1,0 +1,9 @@
+from odoo import models
+
+
+class CommissionSettlement(models.Model):
+    _inherit = "commission.settlement"
+
+    def unlink(self):
+        self.mapped("line_ids.agent_line_partial_ids").unlink()
+        return super().unlink()
