+++ sale_commission_partial_settlement/views/commission_view.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo>
-     <record id="sale_commission_payment_amount" model="ir.ui.view">
-        <field name="name">sale.commission.payment.amount.form</field>
-        <field name="model">sale.commission</field>
-        <field name="inherit_id" ref="sale_commission.sale_commission_form" />
+    <record id="sale_commission_partial_settlement_form_view" model="ir.ui.view">
+        <field name="name">commission.partial.settlement.form.view</field>
+        <field name="model">commission</field>
+        <field name="inherit_id" ref="account_commission.commission_form" />
         <field name="arch" type="xml">
             <field name="invoice_state" position="after">
                 <field
