+++ sale_commission_partial_settlement/wizard/__init__.py
@@ -1,1 +1,1 @@
-from . import wizard_settle
+from . import commission_make_settle
