+++ setup/sale_commission_partial_settlement/odoo/addons/sale_commission_partial_settlement
@@ -1,0 +1,1 @@
+../../../../sale_commission_partial_settlement
